[Clean Architecture] Refactor 1: Introducing "common" Package for Shared Components (#3185)
