+++ app/src/main/java/org/groundplatform/android/common/Constants.kt
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.android
+package org.groundplatform.android.common
 
 import android.content.Context
+import org.groundplatform.android.BuildConfig
 import org.groundplatform.android.ui.map.gms.mog.MogSource
 
-/** Application configuration. */
-object Config {
+/** Application constants. */
+object Constants {
   // Shared preferences.
   const val SHARED_PREFS_NAME = "shared_prefs"
   const val SHARED_PREFS_MODE = Context.MODE_PRIVATE
@@ -55,13 +56,13 @@ object Config {
   /** Limit on the permitted character length for free text question responses. */
   const val TEXT_DATA_CHAR_LIMIT = 255
 
-  // TODO: Make sub-paths configurable and
-  //  stop hardcoding here.
+  // TODO: Make sub-paths configurable and stop hardcoding here.
   // Issue URL: https://github.com/google/ground-android/issues/1730
   const val DEFAULT_MOG_TILE_LOCATION = "/offline-imagery/default"
   private const val DEFAULT_MOG_MIN_ZOOM = 8
   private const val DEFAULT_MOG_MAX_ZOOM = 14
 
+  // TODO: Move to a util class
   fun getMogSources(path: String) =
     listOf(
       MogSource(0..<DEFAULT_MOG_MIN_ZOOM, "$path/$DEFAULT_MOG_MIN_ZOOM/overview.tif"),
@@ -71,5 +72,6 @@ object Config {
       ),
     )
 
+  // TODO: Move to a util class
   fun isReleaseBuild(): Boolean = BuildConfig.BUILD_TYPE.contentEquals("release")
 }
