+++ app/src/main/java/org/groundplatform/android/ui/common/BaseMapViewModel.kt
@@ -44,8 +44,8 @@ import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.flow.withIndex
 import kotlinx.coroutines.launch
-import org.groundplatform.android.Config.DEFAULT_LOI_ZOOM_LEVEL
 import org.groundplatform.android.R
+import org.groundplatform.android.common.Constants.DEFAULT_LOI_ZOOM_LEVEL
 import org.groundplatform.android.model.Survey
 import org.groundplatform.android.model.geometry.Coordinates
 import org.groundplatform.android.model.imagery.TileSource
