+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/multiplechoice/MultipleChoiceItemView.kt
@@ -42,9 +42,9 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import org.groundplatform.android.Config
 import org.groundplatform.android.ExcludeFromJacocoGeneratedReport
 import org.groundplatform.android.R
+import org.groundplatform.android.common.Constants
 import org.groundplatform.android.model.task.MultipleChoice
 import org.groundplatform.android.model.task.Option
 import org.groundplatform.android.ui.theme.AppTheme
@@ -130,10 +130,13 @@ private fun OtherTextField(
     supportingText = {
       Row(modifier = modifier.fillMaxWidth()) {
         Spacer(modifier = modifier.weight(1f))
-        Text("${item.otherText.length} / ${Config.TEXT_DATA_CHAR_LIMIT}", textAlign = TextAlign.End)
+        Text(
+          "${item.otherText.length} / ${Constants.TEXT_DATA_CHAR_LIMIT}",
+          textAlign = TextAlign.End,
+        )
       }
     },
-    isError = item.otherText.length > Config.TEXT_DATA_CHAR_LIMIT,
+    isError = item.otherText.length > Constants.TEXT_DATA_CHAR_LIMIT,
     value = item.otherText,
     textStyle = MaterialTheme.typography.bodyLarge,
     onValueChange = { otherValueChanged(it) },
