+++ app/src/main/java/org/groundplatform/android/ui/map/gms/features/FeatureClusterRenderer.kt
@@ -23,7 +23,7 @@ import com.google.android.gms.maps.model.MarkerOptions
 import com.google.maps.android.clustering.Cluster
 import com.google.maps.android.clustering.ClusterManager
 import com.google.maps.android.clustering.view.DefaultClusterRenderer
-import org.groundplatform.android.Config.CLUSTERING_ZOOM_THRESHOLD
+import org.groundplatform.android.common.Constants.CLUSTERING_ZOOM_THRESHOLD
 import org.groundplatform.android.ui.IconFactory
 import org.groundplatform.android.ui.map.Feature
 import org.groundplatform.android.ui.map.gms.CLUSTER_Z
