+++ app/src/test/java/org/groundplatform/android/ui/datacollection/tasks/text/TextTaskFragmentTest.kt
@@ -20,8 +20,8 @@ import com.google.common.truth.Truth.assertThat
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
-import org.groundplatform.android.Config
 import org.groundplatform.android.R
+import org.groundplatform.android.common.Constants
 import org.groundplatform.android.model.job.Job
 import org.groundplatform.android.model.submission.TextTaskData
 import org.groundplatform.android.model.task.Task
@@ -87,7 +87,7 @@ class TextTaskFragmentTest : BaseTaskFragmentTest<TextTaskFragment, TextTaskView
   fun `text over the character limit is invalid`() = runWithTestDispatcher {
     setupTaskFragment<TextTaskFragment>(job, task)
 
-    runner().inputText("a".repeat(Config.TEXT_DATA_CHAR_LIMIT + 1))
+    runner().inputText("a".repeat(Constants.TEXT_DATA_CHAR_LIMIT + 1))
     // TODO: We should actually validate that the error toast is displayed after Next is clicked.
     // Unfortunately, matching toasts with espresso is not straightforward, so we leave it at
     // an explicit validation check for now.
